fix(game): align event consumer compilation

+++ feature/game/src/main/java/com/adsamcik/tracker/game/event/GameDomainEventConsumer.kt
@@ -105,7 +105,7 @@ class GameDomainEventConsumer @Inject constructor(
 	}
 
 	private suspend fun onDailySummaryUpdated(event: DomainEvent.DailySummaryUpdated) {
-		val cumulativeSteps = event.totalSteps.raw.toInt().coerceAtLeast(0)
+		val cumulativeSteps = event.totalSteps.raw.coerceAtLeast(0)
 		GoalTracker.updateUnqualifiedCumulativeStepsPresentation(cumulativeSteps)
 	}
